Refactored the code to appropriate places
