+++ src/main/java/com/example/demo/config/ApplicationConfig.java
@@ -1,6 +1,6 @@
 package com.example.demo.config;
 
-import com.example.demo.security.UserRepository;
+import com.example.demo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
