+++ src/main/java/com/example/demo/entity/User.java
@@ -1,12 +1,12 @@
-package com.example.demo.security;
+package com.example.demo.entity;
 
+import com.example.demo.util.Role;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
