+++ src/main/java/com/example/demo/filter/JwtAuthenticationFilter.java
@@ -1,5 +1,6 @@
-package com.example.demo.config;
+package com.example.demo.filter;
 
+import com.example.demo.service.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
