+++ src/main/java/com/example/demo/model/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.example.demo.models.request;
+package com.example.demo.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
