+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.security;
+package com.example.demo.repository;
 
-import org.springframework.context.annotation.Scope;
+import com.example.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
