+++ src/main/java/com/example/demo/service/JwtService.java
@@ -1,4 +1,4 @@
-package com.example.demo.config;
+package com.example.demo.service;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
