+++ src/main/java/com/example/demo/util/Role.java
@@ -1,4 +1,4 @@
-package com.example.demo.security;
+package com.example.demo.util;
 
 public enum Role {
     USER,
